backend/src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Dorian <dorian.gloinec@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.2.0"
djangorestframework = "^3.15.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
flake8 = "^7.1.0"
autopep8 = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
